+++ pkg/accounts/service/edit.test.ts
@@ -2,7 +2,7 @@ import { Option, Result } from '@mikuroxina/mini-fn';
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 
 import { Argon2idPasswordEncoder } from '../../password/mod.js';
-import { InMemoryAccountRepository } from '../adaptor/repository/dummy.js';
+import { InMemoryAccountRepository } from '../adaptor/repository/dummy/account.js';
 import { Account, type AccountID } from '../model/account.js';
 import { EditService } from './edit.js';
 import { EtagService } from './etagService.js';
